\name{bhr2000}
\docType{data}
\alias{bhr2000}
\title{Data from Bliese, Halverson and Rothberg (2000)}
\description{
This data set contains the complete data used in Bliese, Halverson & Rotheberg (2000).
The data set contains 14 variables with individual ratings of US Army Company leadership,
work hours, and the degree to which individuals find comfort from religion.  The leadership and
workhours variables are subsets of the Bliese and Halveson (1996) data set; however,
in the case of leadership, the agree data set contains the 11 items that make up the scale
whereas the bh1996 data set contains only the scale score.  Most items are on a strongly
disagree to strongly agree scale.  The RELIG item is on a never to always scale.
}
\usage{data(bhr2000)}
\format{A data frame with 14 columns and 5,400 observations from 99 groups
\tabular{llll}{
[,1] \tab GRP     \tab numeric  \tab Group Identifier\cr
[,2] \tab AF06   \tab numeric  \tab Officers get willing and whole-hearted cooperation\cr
[,3] \tab AF07 \tab numeric  \tab NCOS most always get willing and whole-hearted cooperation\cr
[,4] \tab AP12 \tab numeric  \tab I am impressed by the quality of leadership in this company\cr
[,5] \tab AP17   \tab numeric  \tab I would go for help with a personal problem to the chain of command\cr
[,6] \tab AP33  \tab numeric  \tab Officers in this Company would lead well in combat\cr
[,7] \tab AP34  \tab numeric  \tab NCOs in this Company would lead well in combat\cr
[,8] \tab AS14     \tab numeric  \tab My officers are interested in my personal welfare\cr
[,9] \tab AS15 \tab numeric  \tab My NCOs are interested in my personal welfare\cr
[,10] \tab AS16 \tab numeric  \tab My officers are interested in what I think and feel about things\cr
[,11] \tab AS17 \tab numeric  \tab My NCOs are intested in what I think and fell about things\cr
[,12] \tab AS28 \tab numeric  \tab My chain-of-command works well\cr
[,13] \tab HRS \tab numeric  \tab How many hours do you usually work in a day\cr
[,14] \tab RELIG \tab numeric  \tab How often do you gain strength of comfort from religious beliefs
}
}
\references{
Bliese, P. D. & Halverson, R. R. (1996).  Individual and nomothetic
models of job stress:  An examination of work hours, cohesion, and 
well-being.  Journal of Applied Social Psychology, 26, 1171-1189.

Bliese, P. D., Halverson, R. R., & Rothberg, J. (2000).  Using random group resampling (RGR)
to estimate within-group agreement with examples using the statistical language R.
}
\keyword{datasets}