\name{cohesion}
\docType{data}
\alias{cohesion}
\title{Five cohesion ratings from 11 individuals nested in 4 platoons in 2 larger units}
\description{
This data set contains five cohesion measures provided by 11 individuals.  The individuals
providing the measures are members of four platoons further nested within two larger units.  This
data file is used for demonstative purposes in the document "Multilevel Modeling in R" that
accompanies this package.
}
\usage{data(cohesion)}
\format{A data frame with 7 columns and 11 observations
\tabular{llll}{
[,1] \tab UNIT    \tab numeric  \tab Higher-level Unit Identifier\cr
[,2] \tab PLATOON \tab numeric  \tab Lower-level Platoon Identifier\cr
[,3] \tab COH01   \tab numeric  \tab First Cohesion Variable\cr
[,4] \tab COH02   \tab numeric  \tab Second Cohesion Variable\cr
[,5] \tab COH03   \tab numeric  \tab Third Cohesion Variable\cr
[,6] \tab COH04   \tab numeric  \tab Fourth Cohesion Variable\cr
[,7] \tab COH05   \tab numeric  \tab Fifth Cohesion Variable\cr
}
}
\keyword{datasets}