\name{klein2000}
\docType{data}
\alias{klein2000}
\title{Data from Klein, Bliese, Kozlowski et al., (2000)}
\description{
This data set contains the complete data used in Klein et al. (2000).  Klein et al. (2000)
is a chapter from an edited volume by K. J. Klein and S. W. J. Kozlowski.  In this 
chapter the authors use a simulated data set to compare and contrast WABA, HLM, and
Cross-Level Operator Analyses (CLOP).  The simulated data set was created by Paul Bliese.
}
\usage{data(klein2000)}
\format{A data frame with 9 columns and 750 observations from 50 groups
\tabular{llll}{
[,1] \tab GRPID   \tab numeric  \tab Group Identifier\cr
[,2] \tab JOBSAT  \tab numeric  \tab Job Satisfaction (DV)\cr
[,3] \tab COHES   \tab numeric  \tab Cohesion\cr
[,4] \tab POSAFF  \tab numeric  \tab Positive Affect\cr
[,5] \tab PAY     \tab numeric  \tab Pay \cr
[,6] \tab NEGLEAD \tab numeric  \tab Negative Leadership\cr
[,7] \tab WLOAD   \tab numeric  \tab Workload\cr
[,8] \tab TASKSIG \tab numeric  \tab Task Significance\cr
[,9] \tab PHYSEN  \tab numeric  \tab Physical Environment
}
}
\references{
Klein, K. J., Bliese, P.D., Kozlowski, S. W. J, Dansereau, F., Gavin, M. B.,
Griffin, M. A., Hofmann, D. A., James, L. R., Yammarino, F. J., & Bligh, M. C. (2000).
Multilevel analytical techniques:  Commonalities, differences, and continuing questions.
In K. J. Klein & S. W. Kozlowski (Eds.), Multilevel Theory, Research, and Methods in
Organizations (pp. 512-553). San Francisco, CA:  Jossey-Bass, Inc}
\keyword{datasets}