\name{lq2002}
\docType{data}
\alias{lq2002}
\title{Data used in special issue of Leadership Quarterly, Vol. 13, 2002}
\description{
Contains the complete data used in a special issue of Leadership Quarterly edited
by Paul Bliese, Ronald Halverson and Chet Schriesheim in 2002 (Vol 13). Researchers from several
universities analyzed this common dataset using various multilevel techniques. The three
scales used in the analyses are Leadership Climate (LEAD), Task Significance (TSIG) and Hostility
(HOSTILE). The data set contains each item making up these scales. These items were additionally used by
Cohen, Doveh and Nahum-Shani (2009).
}
\usage{data(lq2002)}

\format{A data frame with 27 columns and 2,042 observations from 49 groups
\tabular{llll}{
[,1] \tab COMPID  \tab numeric  \tab Army Company Identifying Variable\cr
[,2] \tab SUB     \tab numeric  \tab Subject Number\cr
[,3] \tab LEAD01  \tab numeric  \tab Officers Get Cooperation From Company (EXV01)\cr
[,4] \tab LEAD02  \tab numeric  \tab NCOs Get Cooperation From Company (EXV02)\cr
[,5] \tab LEAD03  \tab numeric  \tab Impressed By Leadership (EXV04)\cr
[,6] \tab LEAD04  \tab numeric  \tab Go For Help Within Chain of Command (EXV05)\cr
[,7] \tab LEAD05  \tab numeric  \tab Officers Would Lead Well In Combat (EXV07)\cr
[,8] \tab LEAD06  \tab numeric  \tab NCOs Would Lead Well In Combat (EXV08)\cr
[,9] \tab LEAD07  \tab numeric  \tab Officers Interested In Welfare (EXV11)\cr
[,10] \tab LEAD08 \tab numeric  \tab NCOs Interested In Welfare (EXV13)\cr
[,11] \tab LEAD09 \tab numeric  \tab Officers Interested In What I Think (EXV14)\cr
[,12] \tab LEAD10 \tab numeric  \tab NCOs Interested In What I Think (EXV15)\cr
[,13] \tab LEAD11 \tab numeric  \tab Chain Of Command Works Well (EXV16)\cr
[,14] \tab TSIG01 \tab numeric  \tab What I Am Doing Is Important (MIS05)\cr
[,15] \tab TSIG02 \tab numeric  \tab Making Contribution To Mission (MIS06)\cr
[,16] \tab TSIG03 \tab numeric  \tab What I Am Doing Accomplishes Mission (MIS07)\cr
[,17] \tab HOSTIL01 \tab numeric  \tab Easily Annoyed Or Irritated (BSI09)\cr
[,18] \tab HOSTIL02 \tab numeric  \tab Temper Outburst That You Cannot Control (BSI18)\cr
[,19] \tab HOSTIL03 \tab numeric  \tab Urges To Harm Someone (BSI47)\cr
[,20] \tab HOSTIL04 \tab numeric  \tab Urges To Break Things (BSI49)\cr
[,21] \tab HOSTIL05 \tab numeric  \tab Getting Into Frequent Arguments (BSI54)\cr
[,22] \tab LEAD \tab numeric  \tab Leadership Climate Scale Score\cr
[,23] \tab TSIG \tab numeric  \tab Task Significance Scale Score\cr
[,24] \tab HOSTILE \tab numeric  \tab Hostility Scale Score\cr
[,25] \tab GLEAD \tab numeric  \tab Leadership Climate Scale Score Aggregated By Company\cr
[,26] \tab GTSIG \tab numeric  \tab Task Significance Scale Score Aggregated By Company\cr
[,27] \tab GHOSTILE \tab numeric  \tab Hostility Scale Score Aggregated By Company
}
}
\references{
Bliese, P. D., & Halverson, R. R. (2002).  Using random group resampling in multilevel
research.  Leadership Quarterly, 13, 53-68.

Bliese, P. D., Halverson, R. R., & Schriesheim, C. A. (2002).  Benchmarking multilevel
methods:  Comparing HLM, WABA, SEM, and RGR.  Leadership Quarterly, 13, 3-14.

Cohen, A., Doveh, E., & Nahum-Shani, I. (2009).  Testing agreement for multi-item scales
with the indices rwg(j) and adm(j).  Organizational Research Methods, 12, 148-164.
}
\keyword{datasets}
