\name{tankdat}
\docType{data}
\alias{tankdat}
\title{Tank data from Bliese and Lang (in press)}
\description{
This data set is a partial sample of data collected by Lang and reported in Lang and
Bliese (2009). The tankdat sub-sample was used as an example of discontinuous
growth modeling in Bliese and Lang (in press). The data set is in long (univariate)
format, and contains performance data from 184 participants over 12 repeated
measures on a complex tank simulation task. In the research paradigm, the task
was unexpectedly changed after the first six performance episodes. Discontinuous
growth models were used to examine participants' reactions to the
unexpected change. The data set contains the person-level predictor of
conscientiousness. 
}
\usage{data(tankdat)}
\format{A dataframe with 4 columns and 2208 observations
\tabular{llll}{
[,1] \tab ID       \tab numeric  \tab Participant ID\cr
[,2] \tab CONSC    \tab numeric  \tab Participant Conscientiousness\cr
[,3] \tab TIME    \tab numeric  \tab Time\cr
[,4] \tab SCORE    \tab numeric  \tab Task Performance\cr
}
}
\references{
Bliese, P. D., & Lang, J. W. B. (in press).  Understanding relative and absolute change in
discontinuous growth models: Coding alternatives and implications for hypothesis testing.
Organizational Research Methods.

Lang, J. W. B., & Bliese, P. D. (2009).  General mental ability and two types of adaptation
to unforeseen change:  Applying discontinuous growth models to the task-change paradigm.
Journal of Applied Psychology, 92, 411-428.
}
\keyword{datasets}