\name{univbct}
\docType{data}
\alias{univbct}
\title{Data from Bliese and Ployhart (2002)}
\description{Contains the complete data set used in Bliese and Ployhart (2002) focused on Job
Satisfation. The data is in univariate (i.e., stacked) form. Data were collected at three time points.
}
\usage{data(univbct)}
\format{A data frame with 22 columns and 1485 observations from 495 individuals
\tabular{llll}{
[,1] \tab BTN     \tab numeric  \tab BN Id\cr
[,2] \tab COMPANY   \tab numeric  \tab Co Id\cr
[,3] \tab MARITAL \tab numeric  \tab Marital Status (1 single; 2 married; 3 separated; 4 divorced; 5 other)\cr
[,4] \tab GENDER \tab numeric  \tab Gender (1 male; 2 female)\cr
[,5] \tab HOWLONG    \tab numeric  \tab Time in Unit \cr
[,6] \tab RANK  \tab numeric  \tab Rank\cr
[,7] \tab EDUCATE  \tab numeric  \tab Education\cr
[,8] \tab AGE     \tab numeric  \tab Age\cr
[,9] \tab JOBSAT1   \tab numeric  \tab JOBSAT Time 1\cr
[,10] \tab COMMIT1  \tab numeric  \tab Commitment Time 1\cr
[,11] \tab READY1 \tab numeric  \tab Readiness Time 1\cr
[,12] \tab JOBSAT2   \tab numeric  \tab JOBSAT Time 2\cr
[,13] \tab COMMIT2  \tab numeric  \tab Commitment Time 2\cr
[,14] \tab READY2 \tab numeric  \tab Readiness Time 2\cr
[,15] \tab JOBSAT3   \tab numeric  \tab JOBSAT Time 3\cr
[,16] \tab COMMIT3  \tab numeric  \tab Commitment Time 3\cr
[,17] \tab READY3 \tab numeric  \tab Readiness Time 3\cr
[,18] \tab TIME \tab numeric  \tab 0 to 2 time maker\cr
[,19] \tab JSAT \tab numeric  \tab Jobsat in univariate form \cr
[,20] \tab COMMIT \tab numeric  \tab Commitment in univariate form\cr
[,21] \tab READY \tab numeric  \tab Readiness in univariate form \cr
[,22] \tab SUBNUM \tab numeric  \tab Subject number
}
}
\references{
Bliese, P. D., & Ployhart, R. E. (2002). Growth modeling using random coefficient models:  Model
building, testing and illustrations.  Organizational Research Methods, 5, 362-387.}
\keyword{datasets}